Moved the IR type definitions to their own file to keep the mod file clean and easy to read
